Add react router and a custom-themes route whoohoo
