+++ demo/index.html
@@ -6,7 +6,7 @@
   <![endif]-->
   <meta charset="utf-8">
   <meta http-equiv="X-UA-Compatible" content="IE=edge,chrome=1">
-  <title>Demo</title>
+  <title>Demoes!</title>
   <meta name="description" content="">
   <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no">
   <link rel="stylesheet" type="text/css" href="https://cdnjs.cloudflare.com/ajax/libs/normalize/4.1.1/normalize.min.css" />
@@ -16,7 +16,7 @@
     <p class="browsehappy">You are using an <strong>outdated</strong> browser. Please <a href="http://browsehappy.com/">upgrade your browser</a> to improve your experience.</p>
   <![endif]-->
   <div id="content">
-    <h1>If you can see this, something is broken (or JS is not enabled)!!.</h1>
+    <h1><em>daaaanngggggg!</em></h1>
   </div>
   <script type="text-javascript" src="http://localhost:3000/webpack-dev-server.js"></script>
   <script type="text/javascript" src="assets/main.js"></script>
