+++ package.json
@@ -22,21 +22,26 @@
     "docs-build-static": "echo 'Static docs generation for victory-examples is not supported"
   },
   "dependencies": {
-    "builder-victory-component": "^2.1.0",
     "builder": "~2.9.1",
+    "builder-victory-component": "^2.1.0",
     "d3": "^3.5.6",
     "victory-chart": "^7.0.2"
   },
   "devDependencies": {
     "builder-victory-component-dev": "^2.1.0",
     "chai": "^3.2.0",
+    "formidable-landers": "^1.0.2",
+    "history": "~1.17.0",
     "lodash": "^3.10.1",
     "mocha": "^2.2.5",
     "opener": "^1.4.1",
+    "radium": "^0.17.1",
+    "react-router": "^2.4.0",
     "react": "0.14.x",
     "react-addons-test-utils": "^0.14.0",
     "react-dom": "0.14.x",
     "react-hot-loader": "^1.2.8",
+    "scroll-behavior": "^0.3.3",
     "sinon": "^1.16.1",
     "sinon-chai": "^2.8.0"
   }
